+++ framework/contracts/native/version-control/tests/direct_calls.rs
@@ -11,8 +11,7 @@ type AResult = anyhow::Result<()>; // alias for Result<(), anyhow::Error>
 #[test]
 fn instantiate() -> AResult {
     let chain = MockBech32::new("mock");
-    let sender = chain.sender_addr();
-    let deployment = Abstract::deploy_on(chain, sender.to_string())?;
+    let deployment = Abstract::deploy_on_mock(chain.clone())?;
 
     let factory = deployment.module_factory;
     let factory_config = factory.config()?;
@@ -29,8 +28,7 @@ fn instantiate() -> AResult {
 #[test]
 fn caller_must_be_manager() -> AResult {
     let chain = MockBech32::new("mock");
-    let sender = chain.sender_addr();
-    let deployment = Abstract::deploy_on(chain, sender.to_string())?;
+    let deployment = Abstract::deploy_on_mock(chain.clone())?;
 
     let factory = &deployment.module_factory;
     let test_module = ModuleInfo::from_id(