+++ framework/packages/abstract-sdk/src/cw_helpers/mod.rs
@@ -3,7 +3,9 @@ mod cosmwasm_std;
 mod cw_ownable;
 mod cw_storage_plus;
 mod fees;
+mod migrate_instantiate;
 
 pub use cw_clearable::*;
+pub use migrate_instantiate::*;
 
 pub use self::{cosmwasm_std::*, cw_storage_plus::*, fees::*};