+++ framework/packages/abstract-std/src/objects/mod.rs
@@ -3,6 +3,7 @@
 
 pub(crate) mod ans_asset;
 pub mod ans_host;
+pub mod module_factory;
 pub mod storage_namespaces;
 pub mod version_control;